# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  domain.VPC:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      network_id:
        type: string
    type: object
...
  httputil.Meta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
  httputil.Response:
    properties:
      data: {}
      error: {}
      meta:
        $ref: '#/definitions/httputil.Meta'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Mini AWS Compute API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mini AWS API
  version: "1.0"
paths:
...
  /vpcs:
    get:
      description: Gets a list of all existing VPCs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.VPC'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: List all VPCs
      tags:
      - vpcs
    post:
      consumes:
      - application/json
      description: Creates a new virtual private cloud network
      parameters:
      - description: VPC creation request
        in: body
        name: request
        required: true
        schema:
          properties:
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.VPC'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new VPC
      tags:
      - vpcs
  /vpcs/{id}:
    delete:
      description: Removes a VPC network (must be empty of instances)
      parameters:
      - description: VPC ID or Name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a VPC
      tags:
      - vpcs
    get:
      description: Gets detailed information about a specific VPC
      parameters:
      - description: VPC ID or Name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.VPC'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Get VPC details
      tags:
      - vpcs
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"